Add files via upload
Write a program that does the following: It prompts the user with "Enter a number from 1 to 20:\n" (note the newline after the colon), and reads the user's response. If the user's response was, say, 10, the program then prints a short message followed by the first 10 ordinal numbers (see this article), one per line, and then terminates. If the user enters a number that is outside of the 1..20 range, the program should print the error message "Number is not in the range from 1 to 20\n" (note the newline at the end), and terminate.