Add files via upload
This is an exercise that introduces reading and writing octal and hexadecimal integers, as well as formatting spacing/padding. In this context, please note that decimal means a number written in base 10; it does not mean that it has a decimal point and a fractional part! Likewise, octal is base 8, and hexadecimal (or "hex") is base 16.

Write a program that requests six integers ("Enter six integers:\n") (note newline after colon), reads all of them, then prints all of them in the following format: (i) first print a header line as shown in the example below; then (ii) two integers per line, with each integer right-justified in a field of 10 characters, separated by two blank spaces. Each of the integers supplied could be in either decimal, octal or hexadecimal format, and your program should read it correctly. For output, each integer should be printed in decimal format.