Add files via upload
This is an exercise in nested loops and printing characters.

Write a program that repeatedly reads two integers in decimal format, the first being the width (i.e., number of columns) and the second being the height (i.e., number of rows) of a rectangle. The program prints the rectangle using the characters '+' for each corner and '-' or '|' for each line, and '~' for the interior. The numbers given will all be positive and non-zero and less than 100, except that to end the program, a single 0 is provided for the width (and the program immediately terminates without requiring a height to be entered).

Prompt the user to enter two numbers with "Please enter width and height:\n" (note newline at end)
Assume the user will only enter two valid values for width and height, or a single 0 to stop the program. The number entered will always be in decimal format, so "010" should be interpreted as ten not eight. Tip: Check the format specifier for scanf().
The program will print a rectangle with the specified width and height, using the characters specified.
The program repeats, prompting for the next input, until the user enters a single 0, in which case the program prints "End\n" (note the newline) and terminates.
For example, the following is one execution scenario:

Please enter width and height:
5
3
+---+
|~~~|
+---+
Please enter width and height:
6
2
+----+
+----+
Please enter width and height:
2
2
++
++
Please enter width and height:
4
1
+--+
Please enter width and height:
1
1
+
Please enter width and height:
0
End